Update the new explorer advantabes article to fit into the learn section
